detector/requirements.txt
pyyaml==6.0.2
requests==2.32.3
psutil==6.1.1
flask==3.1.0
